Bowl scoring backend service has been implemented.

+++ src/main/java/com/armanfar/bowlscoring/repository/ScoringRepository.java
@@ -3,5 +3,5 @@ package com.armanfar.bowlscoring.repository;
 import com.armanfar.bowlscoring.entity.BowlScore;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ScoringRepository extends JpaRepository<BowlScore, Long> {
+public interface ScoringRepository extends JpaRepository<BowlScore, Integer> {
 }
